created documentation for functions
