+++ functions/getAllItems.ts
@@ -1,5 +1,5 @@
 import sqlite3 from "sqlite3";
-import { itemDbRow } from "./types/interfaces";
+import { itemDbRow } from "../types/interfaces";
 
 const db = new sqlite3.Database("./db/items.db", (error) => {
   if (error) {
@@ -8,6 +8,14 @@ const db = new sqlite3.Database("./db/items.db", (error) => {
   console.log("Connected AllItems Service to items database");
 });
 
+
+
+/**
+ * Selects all items from the items database.
+ * 
+ * @returns a JSON of each item.
+ * 
+ */
 export async function getAllItems(): Promise<itemDbRow[]> {
   return new Promise((resolve, reject) => {
     db.all(
