+++ functions/searchItemName.ts
@@ -1,17 +1,23 @@
 import sqlite3 from "sqlite3";
-import { itemDbRow } from "./types/interfaces";
+import { itemDbRow } from "../types/interfaces";
 
 const db = new sqlite3.Database("./db/items.db", (error) => {
   if (error) {
     console.error(error.message);
   }
-  console.log("Connected searchItemID service to items database");
+  console.log("Connected searchItemsName service to items database");
 });
-
-export async function searchItemID(item_id: string): Promise<itemDbRow> {
+/**
+ * Searches the item database by a name search
+ * 
+ * @param item_name - name of the item search, could be a simple letter or the full name
+ * 
+ * @returns a JSON of several items that could potentially match the item name search
+ */
+export async function searchItemName(item_name: string): Promise<itemDbRow> {
   return new Promise((resolve, reject) => {
-    db.get(
-      "SELECT * FROM items WHERE item_id =" + item_id,
+    db.all(
+      "SELECT * FROM items WHERE item_name LIKE'%" + item_name + "%'",
 
       (error: any, row: itemDbRow) => {
         if (error) {
